requirements.txt
streamlit>=1.35,<2
opencv-python-headless>=4.8
numpy>=1.26
Pillow>=10
matplotlib>=3.8
requests>=2.32